WIP Setup docker.
Cloning the code is yet to be done.  Using Docker seems to simplify
setting up development environments, but requires users to install
Docker and learn how to manage containers.  This trade-off may not
worthwhile.  Github Codespaces seems to be an easier way to solve the
same problem (setting up consistent development environments for new
contributors).